renamed module to generic pyproject.py
