+++ src/check_python_versions/sources/all.py
@@ -1,7 +1,7 @@
 from .appveyor import Appveyor
 from .github import GitHubActions
 from .manylinux import Manylinux
-from .poetry_pyproject import PoetryPyProject, PoetryPyProjectPythonRequires
+from .pyproject import PoetryPyProject, PoetryPyProjectPythonRequires
 from .setup_py import SetupClassifiers, SetupPythonRequires
 from .tox import Tox
 from .travis import Travis
